drone_landing_system/requirements.txt
tensorflow>=2.8.0
keras>=2.8.0
keras-segmentation>=0.3.0
torch>=1.12.0
torchvision>=0.13.0
opencv-python>=4.5.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
Pillow>=8.3.0
tqdm>=4.62.0
pyyaml>=5.4.0
streamlit>=1.10.0
plotly>=5.3.0
scipy>=1.7.0
albumentations>=1.1.0
segmentation-models>=1.0.1
segmentation-models-pytorch>=0.3.0